update gitignore :wrench

+++ src/providers/Google/index.tsx
@@ -3,19 +3,11 @@ import auth from '@react-native-firebase/auth';
 import {Alert} from 'react-native';
 import {GoogleSignin, statusCodes} from '@react-native-community/google-signin';
 import {UserContext} from '../../App';
-import {Provider, ProviderButton} from '../../components';
+import {ProviderButton} from '../../components';
 import {getProviderButtonTitle} from '../../utils/helpers';
-import {View, Text} from 'react-native';
 import {GOOGLE_WEB_CLIENT_ID} from 'react-native-dotenv';
 
 const PROVIDER_ID = 'google.com';
-
-GoogleSignin.configure({
-  scopes: ['profile', 'email'], // what API you want to access on behalf of the user, default is email and profile
-  webClientId: GOOGLE_WEB_CLIENT_ID, // client ID of type WEB for your server (needed to verify user ID and offline access)
-  //     offlineAccess: true, // if you want to access Google API on behalf of the user FROM YOUR SERVER
-  //      hostedDomain: '', // specifies a hosted domain restriction
-});
 
 function Google() {
   // declare state variables
@@ -26,6 +18,17 @@ function Google() {
     PROVIDER_ID,
   );
 
+  // Effect handlers
+  useEffect(() => {
+    return () => {
+      GoogleSignin.configure({
+        scopes: ['profile', 'email'], // what API you want to access on behalf of the user, default is email and profile
+        webClientId: GOOGLE_WEB_CLIENT_ID, // client ID of type WEB for your server (needed to verify user ID and offline access)
+        //     offlineAccess: true, // if you want to access Google API on behalf of the user FROM YOUR SERVER
+        //      hostedDomain: '', // specifies a hosted domain restriction
+      });
+    };
+  }, []);
   const handleGoogle = async () => {
     if (!loading) {
       setLoading(true);
@@ -33,23 +36,27 @@ function Google() {
       try {
         await GoogleSignin.hasPlayServices();
 
-        if (variant === 'UNLINK') {
-          await user.unlink(PROVIDER_ID);
+        if (user) {
+          if (variant === 'UNLINK') {
+            await user.unlink(PROVIDER_ID);
+          } else if (variant === 'LINK') {
+            const {accessToken, idToken} = await GoogleSignin.getTokens();
+            const credential = auth.GoogleAuthProvider.credential(
+              idToken,
+              accessToken,
+            );
+            await user.linkWithCredential(credential);
+          }
         } else {
-          //  const userInfo = await GoogleSignin.signIn();
           await GoogleSignin.signIn();
+
           const {accessToken, idToken} = await GoogleSignin.getTokens();
           const credential = auth.GoogleAuthProvider.credential(
             idToken,
             accessToken,
           );
 
-          if (variant === 'LINK') {
-            console.log('user wants to link Google')
-            await user.linkWithCredential(credential);
-          } else if (variant === 'SIGN_IN') {
-            await auth().signInWithCredential(credential);
-          }
+          await auth().signInWithCredential(credential);
         }
       } catch (error) {
         switch (error.code) {
@@ -67,6 +74,10 @@ function Google() {
       }
     }
   };
+
+  if (isOnlyProvider) {
+    return null;
+  }
 
   return (
     <ProviderButton
